--- api/src/main/java/com/buurbak/api/security/data/ConfirmationTokenRepository.java
@@ -1,20 +1,0 @@
-package com.buurbak.api.security.data;
-
-import com.buurbak.api.security.domain.ConfirmationToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Repository
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, UUID> {
-
-    @Transactional
-    @Modifying
-    @Query("UPDATE ConfirmationToken c SET c.confirmedAt = ?2 WHERE c.id = ?1")
-    void updateConfirmedAt(UUID tokenId, LocalDateTime confirmedAt);
-}
