--- api/src/main/java/com/buurbak/api/security/data/RoleRepository.java
@@ -1,8 +1,0 @@
-package com.buurbak.api.security.data;
-
-import com.buurbak.api.security.domain.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, String> {
-    Role findByName(String name);
-}
