--- api/src/main/java/com/buurbak/api/security/data/UserRepository.java
@@ -1,21 +1,0 @@
-package com.buurbak.api.security.data;
-
-import com.buurbak.api.security.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository<T extends User> extends JpaRepository<T, UUID> {
-    Optional<User> findByEmail(String email);
-
-    @Transactional
-    @Modifying
-    @Query("UPDATE User u SET u.enabled = true WHERE u.id = ?1")
-    void enableUserById(UUID id);
-}
