--- api/src/main/java/com/buurbak/api/security/domain/ConfirmationToken.java
@@ -1,37 +1,0 @@
-package com.buurbak.api.security.domain;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-public class ConfirmationToken {
-    @Id
-    @GeneratedValue
-    private UUID id;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false)
-    private LocalDateTime expiresAt;
-
-    private LocalDateTime confirmedAt;
-
-    @ManyToOne
-    @JoinColumn(nullable = false)
-    private User user;
-
-    public ConfirmationToken(LocalDateTime createdAt, LocalDateTime expiresAt, User user) {
-        this.createdAt = createdAt;
-        this.expiresAt = expiresAt;
-        this.user = user;
-    }
-}
