--- api/src/main/java/com/buurbak/api/security/domain/Role.java
@@ -1,17 +1,0 @@
-package com.buurbak.api.security.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Role {
-    @Id
-    private String name;
-}
