--- api/src/main/java/com/buurbak/api/security/domain/User.java
@@ -1,72 +1,0 @@
-package com.buurbak.api.security.domain;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.UUID;
-
-@Entity
-@NoArgsConstructor
-@Data
-@Table(name = "user_table") // "user" is a protected table name in PostgreSQL
-@Inheritance( strategy = InheritanceType.SINGLE_TABLE )
-public class User implements UserDetails {
-    @Id
-    @GeneratedValue
-    private UUID id;
-
-    private String email;
-    private String password;
-
-    @ManyToMany(fetch = FetchType.EAGER)
-    private Collection<Role> roles = new ArrayList<>();
-
-    // defaults for all accounts
-    private boolean enabled = false;
-    private boolean locked = false;
-
-    public User(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public User(String email, String password, Collection<Role> roles) {
-        this.email = email;
-        this.password = password;
-        this.roles = roles;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return roles.stream().map(role -> new SimpleGrantedAuthority(role.getName())).toList();
-    }
-
-    @Override
-    public String getUsername() { return email; }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return !locked;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return enabled;
-    }
-}
