+++ api/src/main/java/com/buurbak/api/security/model/Role.java
@@ -1,0 +1,17 @@
+package com.buurbak.api.security.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Role {
+    @Id
+    private String name;
+}
