+++ api/src/main/java/com/buurbak/api/security/repository/RoleRepository.java
@@ -1,0 +1,8 @@
+package com.buurbak.api.security.repository;
+
+import com.buurbak.api.security.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, String> {
+    Role findByName(String name);
+}
