+++ api/src/main/java/com/buurbak/api/security/repository/UserRepository.java
@@ -1,0 +1,21 @@
+package com.buurbak.api.security.repository;
+
+import com.buurbak.api.security.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface UserRepository<T extends User> extends JpaRepository<T, UUID> {
+    Optional<User> findByEmail(String email);
+
+    @Transactional
+    @Modifying
+    @Query("UPDATE User u SET u.enabled = true WHERE u.id = ?1")
+    void enableUserById(UUID id);
+}
