+++ api/src/main/java/com/buurbak/api/security/service/AuthService.java
@@ -1,0 +1,66 @@
+package com.buurbak.api.security.service;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.buurbak.api.security.model.Role;
+import com.buurbak.api.security.model.User;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import static com.buurbak.api.security.config.AuthenticationConfigConstants.SECRET;
+import static com.buurbak.api.security.config.AuthenticationConfigConstants.TOKEN_PREFIX;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+
+@Service
+@RequiredArgsConstructor
+public class AuthService {
+    private final UserService userService;
+    private final TokenService tokenService;
+
+    public Map<String, String> refreshToken(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        String authorizationHeader = request.getHeader(AUTHORIZATION);
+
+        if (authorizationHeader != null && authorizationHeader.startsWith(TOKEN_PREFIX)) {
+            try {
+                String refresh_token = authorizationHeader.substring(TOKEN_PREFIX.length());
+                Algorithm algorithm = Algorithm.HMAC256(SECRET.getBytes());
+                JWTVerifier verifier = JWT.require(algorithm).build();
+                DecodedJWT decodedjwt = verifier.verify(refresh_token);
+                String username = decodedjwt.getSubject();
+
+                User user = userService.findByUsername(username);
+                String access_token = tokenService.generateAccessToken(request, username, user.getRoles().stream().map(Role::getName).collect(Collectors.toList()));
+
+                Map<String, String> tokens = new HashMap<>();
+                tokens.put("access_token", access_token);
+                tokens.put("refresh_token", refresh_token);
+
+                return tokens;
+            } catch (Exception exception) {
+                response.setHeader("error", exception.getMessage());
+                response.setStatus(FORBIDDEN.value());
+
+                Map<String, String> error = new HashMap<>();
+                error.put("error_message", exception.getMessage());
+
+                response.setContentType(APPLICATION_JSON_VALUE);
+                new ObjectMapper().writeValue(response.getOutputStream(), error);
+            }
+        } else {
+            throw new RuntimeException("refresh token is missing");
+        }
+        return null;
+    }
+}
