+++ api/src/main/java/com/buurbak/api/security/service/RoleService.java
@@ -1,0 +1,17 @@
+package com.buurbak.api.security.service;
+
+import com.buurbak.api.security.repository.RoleRepository;
+import com.buurbak.api.security.model.Role;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class RoleService {
+
+    private RoleRepository roleRepository;
+
+    public Role createRole(String roleName) {
+        return roleRepository.save(new Role(roleName));
+    }
+}
