--- api/src/main/java/com/buurbak/api/trailers/application/TrailerService.java
@@ -1,4 +1,0 @@
-package com.buurbak.api.trailers.application;
-
-public class TrailerService {
-}
