+++ api/src/main/java/com/buurbak/api/trailers/controller/TrailerOfferController.java
@@ -1,0 +1,4 @@
+package com.buurbak.api.trailers.controller;
+
+public class TrailerOfferController {
+}
