--- api/src/main/java/com/buurbak/api/trailers/data/TrailerRepository.java
@@ -1,4 +1,0 @@
-package com.buurbak.api.trailers.data;
-
-public interface TrailerRepository {
-}
