--- api/src/main/java/com/buurbak/api/trailers/domain/Trailer.java
@@ -1,4 +1,0 @@
-package com.buurbak.api.trailers.domain;
-
-public class Trailer {
-}
