--- api/src/main/java/com/buurbak/api/trailers/presentation/TrailerOfferController.java
@@ -1,4 +1,0 @@
-package com.buurbak.api.trailers.presentation;
-
-public class TrailerOfferController {
-}
