+++ api/src/main/java/com/buurbak/api/trailers/repository/TrailerRepository.java
@@ -1,0 +1,4 @@
+package com.buurbak.api.trailers.repository;
+
+public interface TrailerRepository {
+}
