+++ api/src/main/java/com/buurbak/api/trailers/service/TrailerService.java
@@ -1,0 +1,4 @@
+package com.buurbak.api.trailers.service;
+
+public class TrailerService {
+}
