--- api/src/main/java/com/buurbak/api/users/application/CustomerService.java
@@ -1,13 +1,0 @@
-package com.buurbak.api.users.application;
-
-import com.buurbak.api.users.data.CustomerRepository;
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class CustomerService {
-
-    private final CustomerRepository customerRepository;
-
-}
