--- api/src/main/java/com/buurbak/api/users/data/CustomerRepository.java
@@ -1,9 +1,0 @@
-package com.buurbak.api.users.data;
-
-import com.buurbak.api.security.data.UserRepository;
-import com.buurbak.api.users.domain.Customer;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends UserRepository<Customer> {
-}
