+++ api/src/main/java/com/buurbak/api/users/repository/CustomerRepository.java
@@ -1,0 +1,9 @@
+package com.buurbak.api.users.repository;
+
+import com.buurbak.api.security.repository.UserRepository;
+import com.buurbak.api.users.model.Customer;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerRepository extends UserRepository<Customer> {
+}
