+++ api/src/main/java/com/buurbak/api/users/service/CustomerService.java
@@ -1,0 +1,13 @@
+package com.buurbak.api.users.service;
+
+import com.buurbak.api.users.repository.CustomerRepository;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class CustomerService {
+
+    private final CustomerRepository customerRepository;
+
+}
